css styling updated
