+++ public/static/index.html
@@ -11,12 +11,12 @@
 <body>
   <header>
     <h1>Weather Forecast Application</h1>
-    <p>Rexburg, Idaho Current Weather and 3-Day Forecast</p>
+    <p>Current Weather and 3-Day Forecast for Rexburg, Idaho</p>
   </header>
 
   <div id="weather-container">
     <section class="current-weather">
-      <h2>Local Weather Conditions</h2>
+      <h2>Rexburg, Idaho</h2>
       <p><img id="weather-icon" src="https://fakeimg.pl/500x300/fff7f7/8a3939?text=Loading+Image&font=noto-serif&font_size=30" alt="weather icon"></p>
       <p id="weather-temp">&nbsp;</p>
       <p id="weather-desc">&nbsp;</p>
@@ -37,16 +37,8 @@
   </footer>
 
 
-  <script src="/dist/currentweather.js" type="module"></script>
-  <script src="/dist/forecast.js" type="module"></script> 
+  <script src="/dist/weather.js" type="module"></script>
 
-
-
-
-
-
-
-  <!-- <script src="../static/TESTING.js"></script> -->
 </body>
 
 </html>
